Introduce MapEditor for map manipulation methods
