+++ src/main/java/com/fundynamic/d2tm/graphics/Theme.java
@@ -1,7 +1,7 @@
 package com.fundynamic.d2tm.graphics;
 
 
-import com.fundynamic.d2tm.game.map.Map;
+import com.fundynamic.d2tm.game.map.MapEditor;
 import org.newdawn.slick.Image;
 import org.newdawn.slick.SpriteSheet;
 
@@ -13,7 +13,7 @@ public class Theme {
         this.spriteSheet = new SpriteSheet(image, tileWidth, tileHeight);
     }
 
-    public Image getTileImage(int row, Map.TerrainFacing facing) {
+    public Image getTileImage(int row, MapEditor.TerrainFacing facing) {
         int column = facing.ordinal();
         return this.spriteSheet.getSprite(column, row);
     }
