+++ src/test/java/com/fundynamic/d2tm/game/map/renderer/TerrainFacingDeterminerTest.java
@@ -1,6 +1,6 @@
 package com.fundynamic.d2tm.game.map.renderer;
 
-import com.fundynamic.d2tm.game.map.Map;
+import com.fundynamic.d2tm.game.map.MapEditor;
 import junit.framework.Assert;
 import org.junit.Test;
 
@@ -8,37 +8,37 @@ public class TerrainFacingDeterminerTest {
 
     @Test
     public void returnsMiddleWhenNoSameTypeOfNeighbours() throws Exception {
-        Assert.assertEquals(Map.TerrainFacing.MIDDLE, Map.getFacing(false, false, false, false));
+        Assert.assertEquals(MapEditor.TerrainFacing.MIDDLE, MapEditor.getFacing(false, false, false, false));
     }
 
     @Test
     public void returnsTopWhenDifferentTypeAbove() throws Exception {
-        Assert.assertEquals(Map.TerrainFacing.TOP, Map.getFacing(false, true, true, true));
+        Assert.assertEquals(MapEditor.TerrainFacing.TOP, MapEditor.getFacing(false, true, true, true));
     }
 
     @Test
     public void returnsRightWhenDifferentTypeRight() throws Exception {
-        Assert.assertEquals(Map.TerrainFacing.RIGHT, Map.getFacing(true, false, true, true));
+        Assert.assertEquals(MapEditor.TerrainFacing.RIGHT, MapEditor.getFacing(true, false, true, true));
     }
 
     @Test
     public void returnsBottomWhenDifferentTypeBottom() throws Exception {
-        Assert.assertEquals(Map.TerrainFacing.BOTTOM, Map.getFacing(true, true, false, true));
+        Assert.assertEquals(MapEditor.TerrainFacing.BOTTOM, MapEditor.getFacing(true, true, false, true));
     }
 
     @Test
     public void returnsLeftWhenDifferentTypeLeft() throws Exception {
-        Assert.assertEquals(Map.TerrainFacing.LEFT, Map.getFacing(true, true, true, false));
+        Assert.assertEquals(MapEditor.TerrainFacing.LEFT, MapEditor.getFacing(true, true, true, false));
     }
 
     @Test
     public void returnsTopRightWhenSameTypeBottomAndLeft() throws Exception {
-        Assert.assertEquals(Map.TerrainFacing.TOP_RIGHT, Map.getFacing(false, false, true, true));
+        Assert.assertEquals(MapEditor.TerrainFacing.TOP_RIGHT, MapEditor.getFacing(false, false, true, true));
     }
 
     @Test
     public void returnsFullWhenAllSameTypeOfNeighbours() throws Exception {
-        Assert.assertEquals(Map.TerrainFacing.FULL, Map.getFacing(true, true, true, true));
+        Assert.assertEquals(MapEditor.TerrainFacing.FULL, MapEditor.getFacing(true, true, true, true));
     }
 
 }
